# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SVC Ecommerce API",
        "description": "API documentation for SVC Coffee Ecommerce Platform - Complete e-commerce solution with cart, checkout, and order management",
        "contact": {
            "name": "SVC Development Team",
            "email": "usman@workforcecommerce.com"
        },
        "license": {
            "name": "Proprietary",
            "url": "https://workforcecommerce.com/license"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        },
        {
            "url": "https://staging.workforcecommerce.com",
            "description": "Staging Server"
        },
        {
            "url": "https://api.workforcecommerce.com",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/api/cart/coupon": {
            "post": {
                "tags": [
                    "Cart"
                ],
                "summary": "Apply coupon code",
                "description": "Apply a coupon code to cart (Not yet implemented - returns 501)",
                "operationId": "applyCoupon",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "cart_id",
                                    "coupon_code"
                                ],
                                "properties": {
                                    "cart_id": {
                                        "type": "string",
                                        "example": "abc123xyz"
                                    },
                                    "coupon_code": {
                                        "type": "string",
                                        "example": "SUMMER2025"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "501": {
                        "description": "Not implemented",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Coupon system not yet implemented"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "apiKey": {
                "type": "apiKey",
                "description": "Enter your API key",
                "name": "X-API-Key",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "apiKey": []
        }
    ]
}